requirements.txt
certifi==2020.12.5
chardet==4.0.0
idna==2.10
isodate==0.6.0
pyparsing==2.4.7
PyYAML==5.3.1
rdflib==5.0.0
requests==2.25.1
six==1.15.0
SPARQLWrapper==1.8.5
urllib3==1.26.2
